Implement LLVM scoping correctly.
LLVM has three scopes: a context-level scope that contains named
types (which have a %name); a module-level scope that contains named
globals (which have a @name); and a function-level scope that
contains instructions (which have a %name). Names in all three scopes
cannot conflict: either the sigil is different, or the sigil is
the same but the names appear in disjoint lexical positions.

So, remove all the logic responsible for handling nested scopes,
since it is pointless.